Move skip nav link into header
