--- src/App.scss
@@ -1,22 +1,0 @@
-@import './styles/base.scss';
-
-main {
-  flex-grow: 1;
-  background-color: #f9f9f9;
-}
-
-.va-api-skipnav {
-  background-color: #f9f9f9;
-  position: absolute;
-  z-index: 100;
-  top: -100px;
-  
-  &:focus {
-    top: 0;
-    outline-offset: 0;
-  }
-
-  &:hover {
-    background-color: #f9f9f9;
-  }
-}
