+++ src/App.tsx
@@ -3,7 +3,6 @@ import * as React from 'react';
 import classNames from 'classnames';
 import { FlagsProvider } from 'flag';
 import { Route } from 'react-router-dom';
-import { HashLink } from 'react-router-hash-link';
 import { ConnectedRouter } from 'react-router-redux';
 
 import { getDeprecatedFlags } from './apiDefs/deprecated';
@@ -16,7 +15,7 @@ import PageContent from './components/PageContent';
 import { history } from './store';
 
 import 'highlight.js/styles/atom-one-dark-reasonable.css';
-import './App.scss';
+import './styles/base.scss';
 
 class App extends React.Component {
   public render() {
@@ -28,17 +27,12 @@ class App extends React.Component {
       <FlagsProvider flags={appFlags}>
         <ConnectedRouter history={history}>
           <div className="vads-u-display--flex">
-            <div className={classNames("vads-u-display--flex", "vads-u-flex-direction--column", "vads-u-min-height--viewport", "vads-u-width--full")}>
-              <HashLink to="#main"
-                className={classNames(
-                  'va-api-skipnav',
-                  'vads-u-padding-x--2',
-                  'vads-u-padding-y--1',
-                )}
-                onClick={this.handleSkipNavClick}
-              >
-                Skip to main content
-              </HashLink>
+            <div className={classNames(
+              "vads-u-display--flex", 
+              "vads-u-flex-direction--column", 
+              "vads-u-min-height--viewport", 
+              "vads-u-width--full",
+            )}>
               <Header />
               <Route path="/" component={PageContent} />
               <Footer />
@@ -67,15 +61,6 @@ class App extends React.Component {
       show_testing_notice: process.env.REACT_APP_SHOW_TESTING_NOTICE === 'true',
       signups_enabled: process.env.REACT_APP_SIGNUPS_ENABLED !== 'false',
     };
-  }
-
-  // need to manually set focus on navigation to #main, since React Router cancels
-  // native anchor click behavior and react-router-hash-link doesn't handle focus
-  private handleSkipNavClick() {
-    const mainElement: HTMLElement | null = document.querySelector('main');
-    if (mainElement) {
-      mainElement.focus();
-    }
   }
 }
